requirements.txt
customtkinter==5.2.1
darkdetect==0.8.0
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pygame==2.5.2
